+++ src/main/index.ts
@@ -2,6 +2,7 @@ export * from "./activity";
 export * from "./client";
 export * from "./embed";
 export * from "./loader";
+export * from "./logger";
 export * from "./readline";
 export * from "./styles";
 export * from "./time";