+++ templates/hybrid/src/commands/shared/index.ts
@@ -1,0 +1,7 @@
+import { EmbedStyle } from "cocoa-discord-utils";
+
+export const style = new EmbedStyle({
+    author: "invoker",
+    // color: 0xFFFF00,
+    // footer: { text: "message you want to put in footer of embed" },
+});