+++ tests/meta.spec.ts
@@ -1,10 +1,8 @@
-import { assert } from "chai";
-
 import { CocoaBuildTime, CocoaVersion } from "../src/meta";
 
 describe("[meta] Meta Module Test", () => {
     it("Version is Version ??? >Useless Test<", () => {
-        assert.isTrue(CocoaVersion.split(".").length >= 3);
-        assert.isTrue(CocoaBuildTime.split("-").length >= 3);
+        expect(CocoaVersion.split(".").length).toBeGreaterThanOrEqual(3);
+        expect(CocoaBuildTime.split("-").length).toBeGreaterThanOrEqual(3);
     });
 });